+++ CHANGES.rst
@@ -59,6 +59,9 @@ Unreleased
    reddit.
  * **[REDDIT]** Added ``DeprecationWarning`` to :meth:`login` as reddit will
    stop supporting cookie-based authentication on 2015/08/03.
+* **[FEATURE]** Added :meth:`hide` and :meth:`unhide`, which accept up to 50
+   fullnames to be hidden at one time. The appropriate methods in
+   :class:`objects.Hideable` now point here instead.
 
 PRAW 2.1.21
 -----------